requirements.txt
accelerate==0.25.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.60.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
async-lru==2.0.5
asyncer==0.0.8
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.1
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
biopython==1.85
bitsandbytes==0.41.3
black==23.12.1
bleach==6.2.0
blessed==1.21.0
blinker==1.9.0
blis==0.7.11
Brotli==1.1.0
browserbase==1.4.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
cohere==5.16.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
conllu==6.0.0
contourpy==1.3.2
coverage==7.10.1
crewai==0.150.0
crewai-tools==0.58.0
cryptography==41.0.7
cycler==0.12.1
cymem==2.0.11
databricks-cli==0.18.0
dataclasses-json==0.6.7
datasets==2.16.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==6.1.3
docker-pycreds==0.4.0
docstring_parser==0.17.0
dspy==2.6.27
dspy-ai==2.4.13
durationpy==0.10
ecdsa==0.19.1
embedchain==0.1.128
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.116.1
fastavro==1.11.1
fastjsonschema==2.21.1
ffmpy==0.6.1
filelock==3.18.0
filetype==1.2.0
flake8==7.0.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2023.10.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gptcache==0.1.44
gpustat==1.1.1
gradio==5.38.2
gradio_client==1.11.0
greenlet==3.2.3
groovy==0.1.2
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.16.0
h2==4.2.0
hnswlib==0.8.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.10.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
json_repair==0.49.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kubernetes==33.1.0
lancedb==0.24.2
langchain==0.3.27
langchain-cohere==0.3.5
langchain-community==0.3.29
langchain-core==0.3.75
langchain-experimental==0.3.4
langchain-google-genai==2.0.10
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
litellm==1.74.3
loguru==0.7.3
lxml==6.0.0
magicattr==0.1.6
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mem0ai==0.1.115
memory-profiler==0.61.0
mistune==3.1.3
mlflow==2.9.2
mmh3==5.1.0
monotonic==1.6
motor==3.3.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.15
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
nmslib-metabrainz==2.1.3
nodeenv==1.9.1
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-ml-py==12.575.51
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.86.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
optuna==4.3.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.1.4
pandas-stubs==2.3.2.250827
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pathtools==0.1.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
platformdirs==4.3.8
playwright==1.54.0
plotly==6.2.0
pluggy==1.6.0
portalocker==2.7.0
posthog==3.25.0
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.4
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==14.0.2
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pybind11==3.0.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydicom==3.0.1
pydub==0.25.1
pyee==13.0.0
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.6.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.9.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.403
pysbd==0.3.4
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-pptx==1.0.2
pytube==15.0.0
pytz==2023.4
pyvis==0.3.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.15.0
qtconsole==5.6.1
QtPy==2.4.3
querystring-parser==1.2.4
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
ruff==0.12.7
safehttpx==0.1.6
safetensors==0.5.3
schema==0.7.7
scikit-learn==1.3.2
scipy==1.15.3
scispacy==0.5.5
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==5.0.0
sentry-sdk==2.30.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
stagehand==0.5.0
starlette==0.47.2
streamlit==1.47.1
structlog==25.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
textblob==0.19.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.3
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.5.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.46.3
typer==0.16.0
types-python-dateutil==2.9.0.20250516
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.3
uvicorn==0.35.0
waitress==2.1.2
wandb==0.15.12
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
win32_setctime==1.2.0
wrapt==1.17.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
